Refactor interaction transformers
The interaction transformer file was quite long and the test file was
about 1,000 lines so before making any changes the various
interaction transformers have been split out into seperate files
