+++ src/apps/interactions/transformers/index.js
@@ -1,0 +1,15 @@
+const transformInteractionResponseToForm = require('./interaction-response-to-form')
+const transformInteractionToListItem = require('./interaction-to-list-item')
+const transformInteractionFormBodyToApiRequest = require('./interaction-form-body-to-api')
+const transformInteractionResponseToViewRecord = require('./interaction-response-to-view')
+const transformInteractionListItemToHaveUrlPrefix = require('./interaction-list-item-to-have-url')
+const transformServiceDeliveryToAttendeeListItem = require('./service-delivery-to-attendee-list-item')
+
+module.exports = {
+  transformInteractionResponseToForm,
+  transformInteractionToListItem,
+  transformInteractionFormBodyToApiRequest,
+  transformInteractionResponseToViewRecord,
+  transformInteractionListItemToHaveUrlPrefix,
+  transformServiceDeliveryToAttendeeListItem,
+}
